Update fetch_budget.php
This is how I troubleshoot the problem of the data now being able to be fetch. The issue was that the request was trying to pull the tables but I had a missing column. When it tried to pull the data it would not display any. It was fixed once the column was added.     Command I used to see the error on the cli was "php fetch_budget.php"